Better return type for pointer math diffs.

+++ Source/Memory/PtrMath.h
@@ -14,28 +14,28 @@ namespace PtrMath
         address += amount;
         return reinterpret_cast<void*>(address);
     }
-    
+
     inline void* Subtract(void* pointer, int amount)
     {
         uintptr_t address = reinterpret_cast<uintptr_t>(pointer);
         address -= amount;
         return reinterpret_cast<void*>(address);
     }
-    
-    inline int Diff(void* lhs, void* rhs)
+
+    inline ptrdiff_t Diff(void* lhs, void* rhs)
     {
         uintptr_t lhsAddr = reinterpret_cast<uintptr_t>(lhs);
         uintptr_t rhsAddr = reinterpret_cast<uintptr_t>(rhs);
-        return lhsAddr - rhsAddr;
+        return static_cast<ptrdiff_t>(lhsAddr - rhsAddr);
     }
-    
+
     inline bool LessThan(void* lhs, void* rhs)
     {
         uintptr_t lhsAddr = reinterpret_cast<uintptr_t>(lhs);
         uintptr_t rhsAddr = reinterpret_cast<uintptr_t>(rhs);
         return lhsAddr < rhsAddr;
     }
-    
+
     /**
      * Given a pointer, returns the nearest larger pointer whose address is aligned as desired.
      * "Aligned" means that the address is divisible by the alignment value.
@@ -49,7 +49,7 @@ namespace PtrMath
         address = address & (~(alignment - 1));
         return reinterpret_cast<void*>(address);
     }
-    
+
     /**
      * Like Align above, but also allows allocating some extra space to store a header.
      * Headers are required by some types of allocators.
@@ -61,11 +61,11 @@ namespace PtrMath
         // See if meeting alignment requirement will naturally leave enough room for header.
         void* alignedPointer = Align(pointer, alignment);
         if(Diff(alignedPointer, pointer) >= headerSize) { return alignedPointer; }
-        
+
         // Normal alignment doesn't leave enough room for the header, so we'll have to manually add space.
         // Pretend that the address passed in was after some number of bytes for a header.
         uintptr_t address = reinterpret_cast<uintptr_t>(pointer) + headerSize;
-        
+
         // Then, align per usual.
         address += (alignment - 1);
         address = address & (~(alignment - 1));
